Move path functions to litestream package
